+++ src/aplikasiinvestasi/dao/impl/LpjDaoImpl.java
@@ -276,7 +276,7 @@ public class LpjDaoImpl implements LpjDao {
                 kredit.setKodeRekening(obj[0].toString());
                 kredit.setBulan(obj[1].toString());
                 kredit.setTahun(obj[2].toString());
-                kredit.setKredit(Long.parseLong(obj[3].toString()));
+                kredit.setKredit(Double.parseDouble(obj[3].toString()));
                 listCredit.add(kredit);
             }
         }catch(  HibernateException | ExceptionInInitializerError e){
@@ -306,7 +306,7 @@ public class LpjDaoImpl implements LpjDao {
                 kredit.setKodeRekening(obj[0].toString());
                 kredit.setBulan(obj[1].toString());
                 kredit.setTahun(obj[2].toString());
-                kredit.setKredit(Long.parseLong(obj[3].toString()));
+                kredit.setKredit(Double.parseDouble(obj[3].toString()));
                 listCredit.add(kredit);
             }
         }catch(  HibernateException | ExceptionInInitializerError e){
